Fix references Fix typo added sln file
