+++ packages.config
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="EXILEDOFFICIAL" version="8.11.0" targetFramework="net481" />
-  <package id="ExMod.Exiled" version="9.3.0" targetFramework="net48" />
+  <package id="ExMod.Exiled" version="9.5.1" targetFramework="net48" />
 </packages>
